#1 Add `HATEOAS` support for `PATCH` middleware
This commit adds `HATEOAS` support for the `PATCH` middleware in the
`courses` controller. The `PATCH` middleware is used to update a
course partially.

Note: the `HATEOAS` functionality is still under development. It is
most likely subject to change.

Note2: the current implementation of `HATEOAS` does not store the `links`
in the database. This means that the `links` are not persisted.  In case
we desire to persist the `links`, we need to add a new field to the
`Course` model.