#1 Update `HATEOAS` for Course entity
This commit makes some changes to the middleware of the Course entity.
If the links of HATEOAS (in the form of an Object) are combined with the
response (i.e. the Course Object), then the response will be in the
form: `{"course": {}, "links": {}}`.

However, the tests of Postman were defined in a way that they expected
just the response without the additional "course" wrapper. This commit
updates the tests to expect the new format of the response.
Furthermore, middleware that do not implement HATEOAS do too return
the response with the "course" wrapper to support consistency.