#1 Add mechanism to handle undefined routes
This commit makes the changes to the codebase to handle undefined routes.
So, if a user tries to access a route that is not defined, the API will
return a 404 - Not Found response.

Furthermore, it provides such tests in the Postman collection.
Ultimately, the tests for the Employee entity were also added.