#2 Update semantics of `POST` with a relationship
Having made some research, the previous implementation of the `POST` method
with a relationship was not RESTful. The `POST` method is used to create a
resouce and link it to another entity. It should not be the case that the
that merely the relationship is created, but the entity itself is not.
Hence this commit changes the semantics of the `POST` method.

The changes are similarly refelcted in the Postman collection. The
collection is also re-organized to make it more readable.