#5 Update `v1` middleware
This commit updates the `v1` middleware of the API. Such a change was
required based on the changes made in the previous commits; furthermore
the `POST` methods have been updated, so an error is returned when a
resource with a duplicate unique key is attempted to be created.

Note: the `POST` with the relation `/students/:id/courses` has not been
updated properly due to an unknown error.