#8 Migrate basecode to Vue.js 3.x
This commit migrates the basecode to Vue.js 3.x.
Note: the new implementation is not a 1:1 copy of the previous implementation.
The new implementation is a simplified version of the previous implementation.
Anyway, it would have been replaced to suit the needs of the project.
As of now, the basecode is intended to be used as a starting point for the project.