Disable `Boostrap.js`
This commit removes the import of `Bootstrap.js` from `main.js`.
Because we're using `Vue.js` and `Bootstrap.js` is not recommended
to use with such a framework, we're removing it.

This rationale is based on the documentation of `Bootstrap.js`:
> https://getbootstrap.com/docs/5.3/getting-started/javascript/#usage-with-javascript-frameworks

Furthermore, we're not using any of the features provided by `Bootstrap.js`
and we're not planning to use them in the future (to not cause any potential
conflicts with `Vue.js`).