#6 Add `term`, `studyPeriod` attrs to `Course`
This commit adds the `term` and `studyPeriod` attributes to the `Course` model.
The `term` attribute is an integer that represents the term in which the course is given.
The `studyPeriod` attribute is an integer that represents the study period in which the course is given.
Both attributes are required.

Furthermore, the `findOneAndUpdate` method is used in the `PUT` and `PATCH` routes of the `courses` controller,
instead of the manual update of the fields, with the help of `findOne` and `save`, previously used.

Ultimately, the `server.postman_collection.json` file is updated to reflect the changes.