#13 Add `jwt-decode` to `client`
This commit adds the `jwt-decode` package to the `client` project.
This package is used to decode the JWT token that is sent from the server to the client.
This token is used to authenticate the user and to get the user's information. The jsonwebtoken
package is used in the server, because it is node-specific and can't be used in the client.