#13 Update Router
This commit makes changes to the Router of the client to accomodate
the logic of the credential system. It also adds a new route for
the login page. Additionally a mechanism, in an experimental stage,
was added to redirect the user to the login page if they try to
access a page that requires authentication without being logged in.