#13 Rename `Student` entity to `User`
We've made a change in the system and the User entity replaces
the Student entity. This commit renames the Student entity to
User in the codebase (all instances, in all files).

Now, we have obtained the abstraction of the User entity that can be
either: (i) Student, or (ii) Admin. Hence, we can implement the
authentication system.