#13 Add authentication mechanism to middleware
This commit makes use of the JWT authentication mechanism to protect
routes that require authentication. The routes that require
authentication are the ones that modify the database, such as POST,
PUT, PATCH and DELETE. The routes that only read from the database
are not protected.

Note: some GET routes are protected, such as with the user; we don't
want to expose the password of the user, so we protect the route.