Fix the Login/logout msg in the header
This commit adds a simple abstraction of an event bus with `mitt`.
This way, we can propagate signals more easily between components.
The struggle was that the user would need to refresh to see the changes
in the button because of the nature of localStorage. Using a ref,
reactive state, or a computed property would not work because the
component is not re-rendered when the localStorage is changed.