Add component that displays API information
As a testing ground for the API, a component that displays the API
information is added to the home page. This component is only visible
when the API is available and the user clicks on a button to show the
information.

Note: the placement of the component is subject to change.