Extend 'CourseInformationView.vue' functionality
This commit adds functionality regarding fetching employee information given a specific course. Furthermore, the 'CourseInformation.vue' component has been updated to display this information as well as including an initial test of a 'EmployeePopUp.vue' component to see if the database fetching/external prop structure works as intended - so that this can be further improved on in a future commit.