Update axios definitions, add HATEOAS compatibility
This commit updates the axios definitions on the client side of all
the entities to necessitate the changes made by the previous commits
on the graphical user interface. It also adds HATEOAS compatibility
with a reusable function that can be used to perform HTTP requests
based on the method type and URL obtained from responses that contain
HATEOAS links.