Add functionality to handle 'ExpiredTokenError'
This commit updates the frontend to handle the `ExpiredTokenError` that
is raised when the token is expired. The frontend now checks for this
error and if it is raised, it deletes the token from the local storage
and navigates to the login page (to force the user to log in again).