Add creation of employee with a course (admin)
This commit adds the possibility to create an employee and link it to a
course. This is done by adding a new button to the admin view, which
allows the user to toggle the creation of a relationship between an
employee and a course. If the user selects this option, then the
employee will be created and linked to the course.