Add proper support for production mode (server)
This commit adds proper support for production mode. The server now
checks for the environment variable ENV_NODE and sets the cookie
sameSite attribute accordingly. The default value is development.

In main.js, the same check is performed to determine how errors are
displayed.

The reason why we don't use `set NODE_ENV=production` is because
this can conflict with some local configurations and environment.
Basically, the value is set globally and can cause problems, namely
outside the project, when running other Node.js applications.